#43 - Extracting ReviewSender to separate service
