+++ src/main/java/dev/iakunin/codexiabot/codexia/service/ReviewSender.java
@@ -1,0 +1,7 @@
+package dev.iakunin.codexiabot.codexia.service;
+
+import dev.iakunin.codexiabot.codexia.entity.CodexiaReview;
+
+public interface ReviewSender {
+    void send(CodexiaReview review);
+}
